smooth wave refactor inside portals component
